backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
cryptography==41.0.7
PyJWT==2.8.0
django-oauth-toolkit==1.7.1
django-filter==23.3
django-extensions==3.2.3
Pillow==10.4.0
twilio==8.10.0
requests==2.31.0
python-decouple==3.8
django-ratelimit==4.1.0
django-auditlog==2.3.0
django-import-export==3.3.1
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
